requirements.txt
Django==4.2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Pillow==10.0.1
razorpay==1.4.1
requests==2.31.0



